feat(triggerPinOrBiometricAuthentication): support new DEVICE_HAS_NO_AUTHENTICATION result case (#211)
With the new PIN & Biometrics 2.0 project
([PRD](https://confluence.tid.es/pages/viewpage.action?pageId=401905341)),
we rely on device authentication to secure the app instead of having a
custom PIN code.

To adapt the existing method to the new implementation, I'm adding more
information to the README and a new response type.